Ensure imagestream api before watching
